Refactor: Define a mixin and abstract login view
kakao 이외의 다른 소셜 서비스도 사용하기 위해 소셜 로그인의 공통
기능을 캡슐화하는 SocialLoginView를 정의.
SocialLoginView 메서드 중 post()를 제외하고 모두 SocialLoginMixin의
메서드로 이동.
